+++ core/src/sql/time/jiff.rs
@@ -1,16 +1,4 @@
-/**
- * Rust type for
- * [time](https://www.postgresql.org/docs/current/datatype-datetime.html).
- */
-pub type Time = time::Time;
-pub type Timezone = time::UtcOffset;
-/**
- * Rust type for
- * [timetz](https://www.postgresql.org/docs/current/datatype-datetime.html).
- */
-pub type TimeTz = (Time, Timezone);
-
-impl crate::ToSql for Time {
+impl crate::ToSql for jiff::civil::Time {
     fn ty(&self) -> crate::pq::Type {
         crate::pq::types::TIME
     }
@@ -19,7 +7,7 @@ impl crate::ToSql for Time {
      * https://github.com/postgres/postgres/blob/REL_12_0/src/backend/utils/adt/date.c#L1235
      */
     fn to_text(&self) -> crate::Result<Option<String>> {
-        self.to_string().to_text()
+        format!("{self:.0}").to_text()
     }
 
     /*
@@ -39,13 +27,14 @@ impl crate::ToSql for Time {
     }
 }
 
-impl crate::FromSql for Time {
+impl crate::FromSql for jiff::civil::Time {
     /*
      * https://github.com/postgres/postgres/blob/REL_12_0/src/backend/utils/adt/date.c#L1170
      */
     fn from_text(ty: &crate::pq::Type, raw: Option<&str>) -> crate::Result<Self> {
-        let format = time::macros::format_description!("[hour]:[minute]:[second]");
-        Time::parse(crate::from_sql::not_null(raw)?, &format).map_err(|_| Self::error(ty, raw))
+        let s = crate::from_sql::not_null(raw)?;
+
+        s.parse().map_err(|_| Self::error(ty, raw))
     }
 
     /*
@@ -54,12 +43,15 @@ impl crate::FromSql for Time {
     fn from_binary(ty: &crate::pq::Type, raw: Option<&[u8]>) -> crate::Result<Self> {
         let usec = i64::from_binary(ty, raw)?;
 
-        Ok(Time::MIDNIGHT + time::Duration::microseconds(usec))
+        Ok(jiff::civil::Time::MIN + jiff::SignedDuration::from_micros(usec))
     }
 }
 
-impl crate::entity::Simple for Time {}
+impl crate::entity::Simple for jiff::civil::Time {}
 
+type TimeTz = (jiff::civil::Time, jiff::tz::Offset);
+
+#[cfg(feature = "time")]
 impl crate::ToSql for TimeTz {
     fn ty(&self) -> crate::pq::Type {
         crate::pq::types::TIMETZ
@@ -78,12 +70,13 @@ impl crate::ToSql for TimeTz {
     fn to_binary(&self) -> crate::Result<Option<Vec<u8>>> {
         let mut buf = self.0.to_binary()?.unwrap();
 
-        crate::to_sql::write_i32(&mut buf, -self.1.whole_seconds())?;
+        crate::to_sql::write_i32(&mut buf, -self.1.seconds())?;
 
         Ok(Some(buf))
     }
 }
 
+#[cfg(feature = "time")]
 impl crate::FromSql for TimeTz {
     /*
      * https://github.com/postgres/postgres/blob/REL_12_0/src/backend/utils/adt/date.c#L1971
@@ -96,8 +89,7 @@ impl crate::FromSql for TimeTz {
             None => return Err(Self::error(ty, raw)),
         };
 
-        let format = time::macros::format_description!("[hour]:[minute]:[second]");
-        let time = Time::parse(&value[0..x], &format).map_err(|_| Self::error(ty, raw))?;
+        let time = value[0..x].parse().map_err(|_| Self::error(ty, raw))?;
 
         let mut tz = value[x..].replace(':', "");
 
@@ -105,10 +97,9 @@ impl crate::FromSql for TimeTz {
             tz.push_str("00");
         }
 
-        let format = time::macros::format_description!("[offset_hour][offset_minute]");
-        let timezone = Timezone::parse(&tz, &format).map_err(|_| Self::error(ty, raw))?;
+        let offset = parse_offset(&tz).ok_or_else(|| Self::error(ty, raw))?;
 
-        Ok((time, timezone))
+        Ok((time, offset))
     }
 
     /*
@@ -120,36 +111,53 @@ impl crate::FromSql for TimeTz {
         let zone = crate::from_sql::read_i32(&mut buf)?;
 
         Ok((
-            Time::MIDNIGHT + time::Duration::microseconds(time),
-            Timezone::from_whole_seconds(-zone).map_err(|_| Self::error(ty, raw))?,
+            jiff::civil::Time::midnight() + jiff::SignedDuration::from_micros(time),
+            jiff::tz::Offset::from_seconds(-zone).map_err(|_| Self::error(ty, raw))?,
         ))
     }
 }
 
+#[cfg(feature = "time")]
 impl crate::entity::Simple for TimeTz {}
+
+fn parse_offset(s: &str) -> Option<jiff::tz::Offset> {
+    let mut offset = jiff::tz::Offset::ZERO;
+
+    if s.starts_with('-') {
+        offset = offset.negate();
+    }
+    let s = s.trim_start_matches(|c| c == '+' || c == '-');
+
+    let (hours, minutes) = s.split_at(2);
+    let span = jiff::Span::new()
+        .hours(hours.parse::<i64>().ok()?)
+        .minutes(minutes.parse::<i64>().ok()?);
+
+    Some(offset + span)
+}
 
 #[cfg(test)]
 mod test {
     crate::sql_test!(
         time,
-        crate::Time,
+        jiff::civil::Time,
         [
-            ("'00:00:00'", crate::Time::MIDNIGHT),
-            ("'01:02:03'", time::macros::time!(01:02:03)),
+            ("'00:00:00'", jiff::civil::Time::midnight()),
+            ("'01:02:03'", jiff::civil::time(1, 2, 3, 0)),
         ]
     );
 
     crate::sql_test!(
         timetz,
-        crate::TimeTz,
+        (jiff::civil::Time, jiff::tz::Offset),
         [
             (
                 "'00:00:00+0000'",
-                (crate::Time::MIDNIGHT, crate::Timezone::UTC)
+                (jiff::civil::Time::midnight(), jiff::tz::Offset::UTC)
             ),
             (
                 "'01:02:03+0200'",
-                (time::macros::time!(01:02:03), time::macros::offset!(+2))
+                (jiff::civil::time(1, 2, 3, 0), jiff::tz::offset(2),)
             ),
         ]
     );