+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `pg16` feature with new connection options;
 - Foreign types alias;
+- Support [jiff](https://crates.io/crates/jiff) crate for date and time types;
 
 ### Changed
 - `from_sql::read_*` and `from_sql::not_null` public;